use an early return

+++ app/components/FavoriteListPage.tsx
@@ -22,36 +22,35 @@ const FavoriteListPage: FC<FavoriteListPageProps> = async ({
   const data: EntityType = await res.json();
   const items = Array.isArray(data) ? data : data?.data;
 
+  if (!items?.length) {
+    return (
+      <JsonIDEBlock
+        data={items}
+        emptyMessage={`No favorite ${title}`}
+      />
+    );
+  }
+
   return (
     <>
-      {!items?.length && (
-        <JsonIDEBlock
-          data={items}
-          emptyMessage={`No favorite ${title}`}
-        />
-      )}
-      {!!items && items?.length > 0 && (
-        <>
-          <h1 className="py-4 text-2xl font-bold text-amber-300">Favorite {title}</h1>
-          <ScrollArea className="h-80 max-h-[60vh] w-full rounded-md border p-4 bg-gray-800">
-            <ul>
-              {items.map((item) => (
-                <li key={item.id} className="p-4">
-                  <GenericFavoriteButton
-                    id={item.refId}
-                    name={item.name}
-                    whichFavorite={routeBase}
-                    redirectTo={EntityRedirectPath[title]}
-                    method="DELETE"
-                  >
-                    Remove {item.name}
-                  </GenericFavoriteButton>
-                </li>
-              ))}
-            </ul>
-          </ScrollArea>
-        </>
-      )}
+      <h1 className="py-4 text-2xl font-bold text-amber-300">Favorite {title}</h1>
+      <ScrollArea className="h-80 max-h-[60vh] w-full rounded-md border p-4 bg-gray-800">
+        <ul>
+          {items.map((item) => (
+            <li key={item.id} className="p-4">
+              <GenericFavoriteButton
+                id={item.refId}
+                name={item.name}
+                whichFavorite={routeBase}
+                redirectTo={EntityRedirectPath[title]}
+                method="DELETE"
+              >
+                Remove {item.name}
+              </GenericFavoriteButton>
+            </li>
+          ))}
+        </ul>
+      </ScrollArea>
     </>
   );
 };
